requirements.txt
Flask==1.1.4
pyzbar==0.1.9
Pillow==6.2.2
